requirements.txt
Pillow>=9.0.0
torch
torchvision
numpy
opencv-python
tqdm
realesrgan
requests
